+++ app/src/components/market/sections/common_sections/transaction/transaction_details_card/index.tsx
