+++ app/src/components/market/sections/common_sections/transaction/transaction_details_line/index.tsx
