+++ app/src/components/market/sections/market_create/steps/items/funding_and_fee_step.tsx
@@ -42,16 +42,16 @@ import {
 } from '../../../../common/common_styled'
 import { CreateCard } from '../../../../common/create_card'
 import { CurrencySelector } from '../../../../common/currency_selector'
-import { DisplayArbitrator } from '../../../../common/display_arbitrator'
 import { GridTransactionDetails } from '../../../../common/grid_transaction_details'
 import { MarketScale } from '../../../../common/market_scale'
 import { SetAllowance } from '../../../../common/set_allowance'
 import { TradingFeeSelector } from '../../../../common/trading_fee_selector'
-import { TransactionDetailsCard } from '../../../../common/transaction_details_card'
-import { TransactionDetailsLine } from '../../../../common/transaction_details_line'
-import { TransactionDetailsRow, ValueStates } from '../../../../common/transaction_details_row'
 import { VerifiedRow } from '../../../../common/verified_row'
 import { WarningMessage } from '../../../../common/warning_message'
+import { DisplayArbitrator } from '../../../common_sections/arbitrators/display_arbitrator'
+import { TransactionDetailsCard } from '../../../common_sections/transaction/transaction_details_card'
+import { TransactionDetailsLine } from '../../../common_sections/transaction/transaction_details_line'
+import { TransactionDetailsRow, ValueStates } from '../../../common_sections/transaction/transaction_details_row'
 import { Outcome } from '../outcomes'
 
 const CreateCardTop = styled(CreateCard)`
