+++ app/src/components/market/sections/market_sell/market_sell.tsx
@@ -6,6 +6,12 @@ import ReactTooltip from 'react-tooltip'
 import styled from 'styled-components'
 
 import { STANDARD_DECIMALS } from '../../../../common/constants'
+import { BigNumberInputReturn } from '../../../../components/common/form/big_number_input'
+import { WarningMessage } from '../../../../components/market/common/warning_message'
+import {
+  TransactionDetailsRow,
+  ValueStates,
+} from '../../../../components/market/sections/common_sections/transaction/transaction_details_row'
 import {
   useAsyncDerivedValue,
   useCompoundService,
@@ -38,17 +44,14 @@ import {
 import { Button, ButtonContainer } from '../../../button'
 import { ButtonType } from '../../../button/button_styling_types'
 import { BigNumberInput, TextfieldCustomPlaceholder } from '../../../common'
-import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { GenericError } from '../../common/common_styled'
 import { GridTransactionDetails } from '../../common/grid_transaction_details'
 import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { TokenBalance } from '../../common/token_balance'
-import { TransactionDetailsCard } from '../../common/transaction_details_card'
-import { TransactionDetailsLine } from '../../common/transaction_details_line'
-import { TransactionDetailsRow, ValueStates } from '../../common/transaction_details_row'
-import { WarningMessage } from '../../common/warning_message'
 import { OutcomeTable } from '../common_sections/outcome_table'
+import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
+import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
 
 const StyledButtonContainer = styled(ButtonContainer)`
   justify-content: space-between;
